Remove backup files
