+++ src/kernels/jupyter/connection/jupyterRequestCreator.web.ts
@@ -141,7 +141,7 @@ export class JupyterRequestCreator implements IJupyterRequestCreator {
         return JupyterWebSockets.get(id);
     }
 
-    public getFetchMethod(): (input: RequestInfo, init?: RequestInit) => Promise<Response> {
+    public getFetchMethod(): typeof fetch {
         return fetch;
     }
